fix: Winners are those with max points
It is not enough that they have reached the point threshold.

Under the current rules, it should be impossible for more than one player to
reach the threshold simultaneously *and* them having different scores when they do so.
However, it is good to not rely on this, just in case it becomes future
under a future set of rules.