refactor: Extract more common parts between ask_valid_input and async version
I made a change in 1ecea2 that was meant to affect both versions, but I only
changed it in the non-async version.

It's hard to share code between a non-async and an async function.
(The only line that changes is how input is received, `input()` vs `await ainput()`,
but there's no way to make that into a "parameter" since one is async and the
other is not.)