refactor: Don't use f-strings for valid8 help_msg
This is to avoid errors due to formatting twice:
once for the f-string, and another done automatically by valid8 with the given context.
If something is formatted containing a bracket, it will be interpreted
as a format field by the second .format() call.

By only using valid8's help_msg formatting, avoiding f-strings, it is easier
to [remember to] properly escape curly braces.